move detail classes into repos package
