+++ app/src/main/java/com/example/getaaccontributors/feature/detail/repos/presenter/DetailPresenter.kt
@@ -1,4 +1,4 @@
-package com.example.getaaccontributors.feature.detail.presenter
+package com.example.getaaccontributors.feature.detail.repos.presenter
 
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleObserver
@@ -6,7 +6,7 @@ import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.OnLifecycleEvent
 import androidx.paging.CombinedLoadStates
 import androidx.paging.LoadState
-import com.example.getaaccontributors.feature.detail.contract.DetailContract
+import com.example.getaaccontributors.feature.detail.repos.contract.DetailContract
 import com.example.getaaccontributors.model.RepoList
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.MainScope
