+++ app/src/main/res/layout/fragment_detail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".feature.detail.view.DetailFragment">
+    tools:context="com.example.getaaccontributors.feature.detail.repos.view.DetailFragment">
 
     <androidx.core.widget.NestedScrollView
         android:id="@+id/nested_scroll_view"
