+++ app/src/main/res/navigation/nav_graph.xml
@@ -16,7 +16,7 @@
     </fragment>
     <fragment
         android:id="@+id/detailFragment"
-        android:name="com.example.getaaccontributors.feature.detail.view.DetailFragment"
+        android:name="com.example.getaaccontributors.feature.detail.repos.view.DetailFragment"
         android:label="{userName}"
         tools:layout="@layout/fragment_detail">
         <argument
