+++ app/src/test/java/com/example/getaaccontributors/feature/detail/repos/view/DetailViewHolderTest.kt
@@ -1,7 +1,7 @@
-package com.example.getaaccontributors.feature.detail.view
+package com.example.getaaccontributors.feature.detail.repos.view
 
 import android.view.View
-import com.example.getaaccontributors.feature.detail.contract.DetailContract
+import com.example.getaaccontributors.feature.detail.repos.contract.DetailContract
 import com.example.getaaccontributors.model.RepoList
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
